♻️ Refactor: SelectorContainer 컴포넌트화
Related to: #3
